+++ allennlp/data/dataset_readers/dataset_reader.py
@@ -48,15 +48,6 @@ class AllennlpLazyDataset(IterableDataset):
     def index_with(self, vocab: Vocabulary):
         self.vocab = vocab
 
-    def __len__(self):
-        """
-        We rely in a couple of places that calling len on the dataloader
-        (which in turn calls len on the dataset) doesn't raise an error.
-        In the case that you have an IterableDataset and you call len, the pytorch dataloader
-        actually spits out a warning - but we need actually calling it to not crash.
-        """
-        return 1
-
 
 class _LazyInstances(AllennlpLazyDataset):
     """
@@ -122,9 +113,6 @@ class _MaxLazyInstances(AllennlpLazyDataset):
     def index_with(self, vocab: Vocabulary):
         self.inner.index_with(vocab)
 
-    def __len__(self):
-        return len(self.inner)
-
 
 class _DistributedLazyInstances(AllennlpLazyDataset):
     def __init__(self, inner: AllennlpLazyDataset) -> None:
@@ -143,9 +131,6 @@ class _DistributedLazyInstances(AllennlpLazyDataset):
 
     def index_with(self, vocab: Vocabulary):
         self.inner.index_with(vocab)
-
-    def __len__(self):
-        return len(self.inner)
 
 
 class DatasetReader(Registrable):