+++ allennlp/modules/token_embedders/embedding.py
@@ -630,11 +630,10 @@ class EmbeddingsTextFile(Iterator[str]):
         return next(self._iterator)
 
     def __len__(self) -> Optional[int]:
-        """ Hack for tqdm: no need for explicitly passing `total=file.num_tokens` """
         if self.num_tokens:
             return self.num_tokens
         raise AttributeError(
-            'an object of type EmbeddingsTextFile has "len()" only if the underlying '
+            "an object of type EmbeddingsTextFile implements `__len__` only if the underlying "
             "text file declares the number of tokens (i.e. the number of lines following)"
             "in the first line. That is not the case of this particular instance."
         )